pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gestaolegal"
version = "0.1.0"
description = "Sistema de Gestão Legal"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.1",
    "certifi==2019.11.28",
    "email_validator==2.1.0.post1",
    "flask==3.1.0",
    "flask-bcrypt",
    "flask-login",
    "flask-mail",
    "flask-marshmallow",
    "flask-paginate",
    "flask-wtf",
    "gunicorn==20.0.4",
    "pymysql",
    "pytz",
    "jinja2==3.1.6",
    "marshmallow==3.6.1",
    "marshmallow-sqlalchemy==0.23.1",
    "python-dateutil==2.8.1",
    "python-dotenv==1.0.0",
    "python-editor==1.0.4",
    "sqlalchemy==2.0.27",
    "toml==0.10.1",
    "wtforms==3.0.1",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.optional-dependencies]
dev = [
    "pytest-base-url>=2.1.0",
    "pytest-playwright==0.7.0",
    "aiosmtpd==1.4.6",
    "pytest-base-url",
]
